moved interfaces and types to Types folder
