+++ packages/filter/src/lib/types/callback.ts
@@ -1,0 +1,10 @@
+
+export interface Callback<TCallback> {
+    id: string;
+    callback: TCallback;
+}
+
+export interface OnCallbackSet {
+    id: string;
+    unSubscribe: () => void;
+}
