+++ packages/filter/src/lib/types/index.ts
@@ -1,1 +1,4 @@
 export * from './filter';
+export * from './callback';
+export * from './events';
+export * from './search';
